requirements.txt
python-binance==1.0.19
pandas==2.2.0
pandas-ta==0.3.14b
numpy==1.26.0
python-dotenv==1.0.0
loguru==0.7.2
APScheduler==3.10.4
flask==3.1.0
flask-cors==5.0.1

